Prefer cardActionMiddleware in favor of windowOpenPonyfill
